Refactor useRenderTasks hook and add jest for testing
This commit refactors the useRenderTasks hook to organize code better and improve readability. Specifically, the 'findNextUpdateDate' function is refactored in a separate file. Also, tasks can now be grouped into categories in addition to groups.

Furthermore, 'jest' was added as a devDependency and some tests were created. This will help to verify the correctness of the hook's functionality and make it easier to update in the future.

Lastly, the client's package.json is updated to support ES6 modules and change the test script to use Jest.